# Open-SMS/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenSMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface GatewayConfig {
  id: number;
  gatewayUrl: string;
  apiKey: string;
  name: string;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateGatewayConfigBody {
  gatewayUrl: string;
  apiKey: string;
  name: string;
}

export type GatewayHealthStatus =
  (typeof GatewayHealthStatus)[keyof typeof GatewayHealthStatus];

export const GatewayHealthStatus = {
  ok: "ok",
  error: "error",
  unreachable: "unreachable",
} as const;

export interface GatewayHealth {
  status: GatewayHealthStatus;
  /** @nullable */
  uptimeSeconds: number | null;
  /** @nullable */
  queueDepth: number | null;
  /** @nullable */
  smsSentToday: number | null;
  /** @nullable */
  paused: boolean | null;
  /** @nullable */
  version: string | null;
  gatewayUrl: string;
}

/**
 * Template variables
 */
export type SendMessageBodyVars = { [key: string]: string };

export interface SendMessageBody {
  /** Recipient phone number in E.164 format */
  to: string;
  /**
   * Template name to use
   * @nullable
   */
  template?: string | null;
  /** Template variables */
  vars?: SendMessageBodyVars;
  /**
   * Raw message body (use instead of template)
   * @nullable
   */
  body?: string | null;
  /**
   * Webhook URL for delivery receipt
   * @nullable
   */
  callbackUrl?: string | null;
}

export interface SendMessageResponse {
  messageId: string;
  status: string;
  queuedAt: string;
}

export type MessageStatusStatus =
  (typeof MessageStatusStatus)[keyof typeof MessageStatusStatus];

export const MessageStatusStatus = {
  queued: "queued",
  sent: "sent",
  delivered: "delivered",
  failed: "failed",
  pending: "pending",
} as const;

export interface MessageStatus {
  messageId: string;
  to: string;
  /** @nullable */
  template?: string | null;
  status: MessageStatusStatus;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface Template {
  id: number;
  name: string;
  body: string;
  vars: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ListTemplatesResponse {
  templates: Template[];
}

export interface CreateTemplateBody {
  name: string;
  body: string;
}

export interface UpdateTemplateBody {
  body: string;
}

export type MessageLogStatus =
  (typeof MessageLogStatus)[keyof typeof MessageLogStatus];

export const MessageLogStatus = {
  queued: "queued",
  sent: "sent",
  delivered: "delivered",
  failed: "failed",
  pending: "pending",
} as const;

export interface MessageLog {
  id: number;
  messageId: string;
  to: string;
  toMasked: string;
  /** @nullable */
  template?: string | null;
  body: string;
  status: MessageLogStatus;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: string;
}

export interface ListLogsResponse {
  logs: MessageLog[];
  total: number;
}

export interface StatsResponse {
  sentToday: number;
  sentThisWeek: number;
  failedToday: number;
  pending: number;
  delivered: number;
  total: number;
}

export interface Settings {
  autoStart: boolean;
  notifyOnFailure: boolean;
  smsRateLimit: number;
  /** @nullable */
  webhookUrl?: string | null;
}

export interface UpdateSettingsBody {
  autoStart?: boolean;
  notifyOnFailure?: boolean;
  smsRateLimit?: number;
  /** @nullable */
  webhookUrl?: string | null;
}

export type GetLogsParams = {
  /**
   * Filter logs by status
   */
  status?: GetLogsStatus;
  /**
   * Max number of logs to return
   */
  limit?: number;
};

export type GetLogsStatus = (typeof GetLogsStatus)[keyof typeof GetLogsStatus];

export const GetLogsStatus = {
  all: "all",
  delivered: "delivered",
  sent: "sent",
  failed: "failed",
  pending: "pending",
} as const;