# Open-SMS/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenSMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get gateway configuration
 */
export const GetGatewayConfigResponse = zod.object({
  id: zod.number(),
  gatewayUrl: zod.string(),
  apiKey: zod.string(),
  name: zod.string(),
  isDefault: zod.boolean(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update gateway configuration
 */
export const UpdateGatewayConfigBody = zod.object({
  gatewayUrl: zod.string(),
  apiKey: zod.string(),
  name: zod.string(),
});

export const UpdateGatewayConfigResponse = zod.object({
  id: zod.number(),
  gatewayUrl: zod.string(),
  apiKey: zod.string(),
  name: zod.string(),
  isDefault: zod.boolean(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get phone gateway health status
 */
export const GetGatewayHealthResponse = zod.object({
  status: zod.enum(["ok", "error", "unreachable"]),
  uptimeSeconds: zod.number().nullable(),
  queueDepth: zod.number().nullable(),
  smsSentToday: zod.number().nullable(),
  paused: zod.boolean().nullable(),
  version: zod.string().nullable(),
  gatewayUrl: zod.string(),
});

/**
 * @summary Send an SMS message via the gateway
 */
export const SendMessageBody = zod.object({
  to: zod.string().describe("Recipient phone number in E.164 format"),
  template: zod.string().nullish().describe("Template name to use"),
  vars: zod
    .record(zod.string(), zod.string())
    .optional()
    .describe("Template variables"),
  body: zod
    .string()
    .nullish()
    .describe("Raw message body (use instead of template)"),
  callbackUrl: zod
    .string()
    .nullish()
    .describe("Webhook URL for delivery receipt"),
});

/**
 * @summary Get status of a specific message
 */
export const GetMessageStatusParams = zod.object({
  messageId: zod.coerce.string(),
});

export const GetMessageStatusResponse = zod.object({
  messageId: zod.string(),
  to: zod.string(),
  template: zod.string().nullish(),
  status: zod.enum(["queued", "sent", "delivered", "failed", "pending"]),
  sentAt: zod.date().nullish(),
  deliveredAt: zod.date().nullish(),
  error: zod.string().nullish(),
});

/**
 * @summary List all SMS templates
 */
export const ListTemplatesResponse = zod.object({
  templates: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      body: zod.string(),
      vars: zod.array(zod.string()),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
});

/**
 * @summary Create a new SMS template
 */
export const CreateTemplateBody = zod.object({
  name: zod.string(),
  body: zod.string(),
});

/**
 * @summary Get a specific template
 */
export const GetTemplateParams = zod.object({
  name: zod.coerce.string(),
});

export const GetTemplateResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  body: zod.string(),
  vars: zod.array(zod.string()),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a template
 */
export const UpdateTemplateParams = zod.object({
  name: zod.coerce.string(),
});

export const UpdateTemplateBody = zod.object({
  body: zod.string(),
});

export const UpdateTemplateResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  body: zod.string(),
  vars: zod.array(zod.string()),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a template
 */
export const DeleteTemplateParams = zod.object({
  name: zod.coerce.string(),
});

/**
 * @summary Get message logs
 */
export const GetLogsQueryParams = zod.object({
  status: zod
    .enum(["all", "delivered", "sent", "failed", "pending"])
    .optional()
    .describe("Filter logs by status"),
  limit: zod.coerce
    .number()
    .optional()
    .describe("Max number of logs to return"),
});

export const GetLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      id: zod.number(),
      messageId: zod.string(),
      to: zod.string(),
      toMasked: zod.string(),
      template: zod.string().nullish(),
      body: zod.string(),
      status: zod.enum(["queued", "sent", "delivered", "failed", "pending"]),
      sentAt: zod.date().nullish(),
      deliveredAt: zod.date().nullish(),
      error: zod.string().nullish(),
      createdAt: zod.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get message statistics
 */
export const GetStatsResponse = zod.object({
  sentToday: zod.number(),
  sentThisWeek: zod.number(),
  failedToday: zod.number(),
  pending: zod.number(),
  delivered: zod.number(),
  total: zod.number(),
});

/**
 * @summary Get application settings
 */
export const GetSettingsResponse = zod.object({
  autoStart: zod.boolean(),
  notifyOnFailure: zod.boolean(),
  smsRateLimit: zod.number(),
  webhookUrl: zod.string().nullish(),
});

/**
 * @summary Update application settings
 */
export const UpdateSettingsBody = zod.object({
  autoStart: zod.boolean().optional(),
  notifyOnFailure: zod.boolean().optional(),
  smsRateLimit: zod.number().optional(),
  webhookUrl: zod.string().nullish(),
});

export const UpdateSettingsResponse = zod.object({
  autoStart: zod.boolean(),
  notifyOnFailure: zod.boolean(),
  smsRateLimit: zod.number(),
  webhookUrl: zod.string().nullish(),
});