# Open-SMS/lib/api-zod/src/generated/types/gatewayConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenSMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GatewayConfig {
  id: number;
  gatewayUrl: string;
  apiKey: string;
  name: string;
  isDefault: boolean;
  createdAt: Date;
  updatedAt: Date;
}