# Open-SMS/lib/api-zod/src/generated/types/sendMessageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenSMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageResponse {
  messageId: string;
  status: string;
  queuedAt: Date;
}